requirements.txt
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.32.0,<1.0
httpx>=0.27.0,<1.0
pydantic>=2.10.0,<3.0
redis>=5.2.0,<6.0
openai>=1.55.0,<2.0
python-dotenv>=1.0.0,<2.0
